requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.1
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.0
fastapi==0.116.1
fastapi-mail==1.4.1
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-text-splitters==0.3.11
langsmith==0.4.21
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
motor==3.6.0
nest-asyncio==1.6.0
orjson==3.11.3
packaging==25.0
parso==0.8.5
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.6.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
tenacity==9.1.2
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
zstandard==0.24.0